# openapi.yaml
openapi: "3.0.0"
info:
  title: Perfably
  version: 1.0.0
  description: |
    The API documentation of perfably
  contact:
    name: GitHub
    url: https://github.com/FACT-Finder/perfably
    email: ff-dev@fact-finder.com
  license:
    name: "AGPL-3"
    url: "https://github.com/FACT-Finder/perfably/blob/main/LICENSE"
paths:
...
  /project/{project}/report/{version}:
    delete:
      operationId: deleteReport
      summary: Remove a report with all metrics and meta values.
      security:
        - basic: []
      parameters:
        - name: project
          in: path
          description: The project
          required: true
          schema:
            type: string
        - name: version
          in: path
          description: The version of the report
          required: true
          schema:
            type: string
      responses:
        "204":
          description: >-
            The report was deleted.
        default:
          description: >-
            Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: addMetrics
      summary: Add metrics to a specific version
      description: |
        If a metric already exists it will be overridden.
      security:
        - basic: []
      parameters:
        - name: project
          in: path
          description: The project
          required: true
          schema:
            type: string
        - name: version
          in: path
          description: The version of the report
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Metrics"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Metrics"
      responses:
        "204":
          description: >-
            The metrics were successfully added.
        default:
          description: >-
            Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
  schemas:
...
    ApiError:
      type: object
      required:
        - error
        - description
      properties:
        error:
          type: string
          example: Bad Request
        description:
          type: string
          example: "could not parse request: invalid json: EOF"
    Metrics:
      type: object
      additionalProperties:
        type: number
        format: double
      example: >-
        {
          "lines_of_code": 5,
          "parallel_quering_ms(avg)": 100,
          "parallel_quering_ms(99percentile)": 900
        }
...